Replace (unsigned) long -> (u)int32_t

+++ src/CalcManager/CEngine/Number.cpp
@@ -31,7 +31,7 @@ namespace CalcEngine
         PNUMBER ret = _createnum(static_cast<uint32_t>(this->Mantissa().size()) + 1);
         ret->sign = this->Sign();
         ret->exp = this->Exp();
-        ret->cdigit = static_cast<long>(this->Mantissa().size());
+        ret->cdigit = static_cast<int32_t>(this->Mantissa().size());
 
         MANTTYPE *ptrRet = ret->mant;
         for (auto const& digit : this->Mantissa())
