+++ src/CalcManager/Ratpack/basex.cpp
@@ -48,7 +48,7 @@ void __inline mulnumx( PNUMBER *pa, PNUMBER b )
         else
             {
             // if pa is one and b isn't just copy b. and adjust the sign.
-            long sign = (*pa)->sign;
+            int32_t sign = (*pa)->sign;
             DUPNUM(*pa,b);
             (*pa)->sign *= sign;
             }
@@ -86,14 +86,14 @@ void _mulnumx( PNUMBER *pa, PNUMBER b )
     MANTTYPE *ptrc;         // ptrc is a pointer to the mantissa of c.
     MANTTYPE *ptrcoffset;   // ptrcoffset, is the anchor location of the next
                             // single digit multiply partial result.
-    long iadigit=0;         // Index of digit being used in the first number.
-    long ibdigit=0;         // Index of digit being used in the second number.
+    int32_t iadigit=0;         // Index of digit being used in the first number.
+    int32_t ibdigit=0;         // Index of digit being used in the second number.
     MANTTYPE      da=0;     // da is the digit from the fist number.
     TWO_MANTTYPE  cy=0;     // cy is the carry resulting from the addition of
                             // a multiplied row into the result.
     TWO_MANTTYPE  mcy=0;    // mcy is the resultant from a single
                             // multiply, AND the carry of that multiply.
-    long  icdigit=0;        // Index of digit being calculated in final result.
+    int32_t  icdigit=0;        // Index of digit being calculated in final result.
 
     a=*pa;
 
@@ -162,7 +162,7 @@ void _mulnumx( PNUMBER *pa, PNUMBER b )
 //
 //    FUNCTION: numpowlongx
 //
-//    ARGUMENTS: root as number power as long
+//    ARGUMENTS: root as number power as int32_t
 //               number.
 //
 //    RETURN: None root is changed.
@@ -174,7 +174,7 @@ void _mulnumx( PNUMBER *pa, PNUMBER b )
 //
 //-----------------------------------------------------------------------------
 
-void numpowlongx( _Inout_ PNUMBER *proot, _In_ long power )
+void numpowlongx( _Inout_ PNUMBER *proot, _In_ int32_t power )
 
 {
     PNUMBER lret = longtonum( 1, BASEX );
@@ -232,7 +232,7 @@ void __inline divnumx( PNUMBER *pa, PNUMBER b, int32_t precision)
         else
             {
             // if pa is one and b is not one, just copy b, and adjust the sign.
-            long sign = (*pa)->sign;
+            int32_t sign = (*pa)->sign;
             DUPNUM(*pa,b);
             (*pa)->sign *= sign;
             }
@@ -266,10 +266,10 @@ void _divnumx( PNUMBER *pa, PNUMBER b, int32_t precision)
                             // guesses one bit too far.
     PNUMBER tmp = nullptr;     // current guess being worked on for divide.
     PNUMBER rem = nullptr;     // remainder after applying guess.
-    long cdigits;           // count of digits for answer.
+    int32_t cdigits;           // count of digits for answer.
     MANTTYPE *ptrc;         // ptrc is a pointer to the mantissa of c.
 
-    long thismax = precision + g_ratio; // set a maximum number of internal digits
+    int32_t thismax = precision + g_ratio; // set a maximum number of internal digits
                                  // to shoot for in the divide.
 
     a=*pa;
@@ -301,7 +301,7 @@ void _divnumx( PNUMBER *pa, PNUMBER b, int32_t precision)
 
     while ( cdigits++ < thismax && !zernum(rem) )
         {
-        long digit = 0;
+        int32_t digit = 0;
         *ptrc = 0;
         while ( !lessnum( rem, b ) )
             {
